moved rls from utils to models
